requirements.txt
# ============================================
# ONTIX Universal Platform
# Python Dependencies
# ============================================

# ============================================
# Web Framework
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ============================================
# Database & Vector Store
# ============================================
neo4j==5.15.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# ============================================
# LangChain & AI
# ============================================
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-core==0.1.10
openai==1.10.0
tiktoken==0.5.2

# ============================================
# Template & Config
# ============================================
jinja2==3.1.3
pyyaml==6.0.1
python-dotenv==1.0.0
jsonschema==4.20.0

# ============================================
# CLI Tools
# ============================================
typer==0.9.0
rich==13.7.0

# ============================================
# Utilities
# ============================================
requests==2.31.0
httpx==0.26.0

# ============================================
# Data Pipeline & Crawlers
# ============================================
apify-client==1.7.1

# ============================================
# Development Tools
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# ============================================
# Monitoring & Logging
# ============================================
prometheus-client==0.19.0
python-json-logger==2.0.7
